+++ changelog.md
@@ -9,6 +9,7 @@
 - Fixed blueprints not showing up in the creative menu and JEI (Malte)
 - Fixed turrets being rotated incorrectly (Malte)
 - Fixed mixers voiding one tick worth of output (Malte)
+- Fixed screwdriver not being able to sneak+rightclick blocks (BluSunrize)
 
 ##### Version 1.15.2-4.0.0-117
 - Changed rendering of multiblocks to split models into blocks;